Add ParseResult and Parser trait
ParseResult follows the standard Rust pattern of providing a specific
type of Result in a library which encapsulates common usage.

The Parser trait reads better than the repeated long-winded usage
introduced earlier. This then allows us to write much more succict
right and left combinators to take only the side of the pair that we're
interested in.